Compiling some updates from kernel.

+++ src/log.c
@@ -1,2 +1,2 @@
 __thread int mute_flag;
-__thread long unsigned int tid_;
+__thread uint tid_;
